chore(flashblocks): refactor the state processor to simplify
